requirements.txt
pandas
SQLAlchemy
boto3
numpy
scikit-learn>=1.3.2
statsmodels
xgboost
lightgbm
flask_cors

